refactor: rewrite cjsToEsm.js in ECMAScript notation
BREAKING CHANGE: file rename from "cjsToEsm.cjs" to "esm.mjs"
